UNR-5710: Split the startup handler file into a client- and server-specific ones
